Create a jsPsychMonkeys project for your data

run_initial_setup() will read your data and create a jsPsychMonkeys project tailoring the _targets.R file to the tasks included in the data.

run_initial_setup(
  pid,
  download_files = FALSE,
  data_location = NULL,
  download_task_script = FALSE,
  folder = "~/Downloads/jsPsychMonkeystest",
  sensitive_tasks = c(""),
  dont_ask = FALSE,
  open_rstudio = TRUE
)

Arguments

pid

project id

download_files

Download the data files? FALSE / TRUE

If TRUE, requires sFTP server credentials to be located in .vault/credentials

See .vault/credentials_TEMPLATE for more details

data_location

Local folder where the raw data for the project is located

download_task_script

should download the task scripts? (requires server credentials) FALSE / TRUE

folder

location for the project

sensitive_tasks

short names of the sensitive tasks in the protocol, if any

dont_ask

answer YES to all questions so the process runs uninterrupted. This will:

open_rstudio

Open RStudio with the new project TRUE / FALSE

Value

Opens a new RStudio project

Examples

run_initial_setup(pid = 999, download_files = FALSE,
data_location = system.file("extdata", package = "jsPsychMonkeys"),
download_task_script = FALSE, 
folder = tempdir(), 
sensitive_tasks = c(""), dont_ask = TRUE, open_rstudio = FALSE)
#> Error in cli_message(h1_title = "Setup"): could not find function "cli_message"