Process the parameters_monkeys_minimal parameters so a minimal set of input parameters can be used. Also, tries to take into account dependencies between parameters.

Process the parameters_monkeys_minimal parameters so a minimal set of input parameters can be used. Also, tries to take into account dependencies between parameters.

Usage

set_parameters(
  parameters_monkeys_minimal = parameters_monkeys_minimal,
  credentials_folder = ".vault/",
  uid = 1,
  times_repeat_protocol = 1,
  time_to_sleep_before_repeating_protocol = 1,
  browserName = "chrome",
  big_container = FALSE,
  keep_alive = FALSE,
  folder_downloads = NULL,
  DEBUG = FALSE,
  screenshot = FALSE,
  debug_file = FALSE,
  console_logs = FALSE,
  open_VNC = FALSE,
  pid = NULL,
  uid_URL = TRUE,
  local_or_server = NULL,
  local_folder_tasks = NULL,
  server_folder_tasks = NULL,
  disable_web_security = FALSE,
  initial_wait = 2,
  wait_retry = 0.1,
  forced_random_wait = FALSE,
  forced_refresh = FALSE,
  forced_seed = 11
)

Arguments

parameters_monkeys_minimal

A list with the main input parameters. See parameters_monkeys_minimal in _targets.R. If the parameter is entered in this list, use it. You can also enter the parameter directly

credentials_folder

folder where files "SERVER_PATH.R" and ".credentials" are. Usually .vault/

uid

user id's for the monkeys. Can be a single number (e.g. 55) or a numeric vector (e.g. 1:10). Determines the number of monkeys that will be released

times_repeat_protocol

if different than 1, creates an URL parameter ID and multiple links changing the uid

time_to_sleep_before_repeating_protocol

In seconds, how long to sleep before repeating protocol

browserName

In which browser should the monkey run the protocol? c("chrome", "firefox")

big_container

Big containers are needed for big protocols: FALSE / TRUE

keep_alive

Keep the docker container alive after finishing the protocol? This is useful to debug.

folder_downloads

Local folder where csv's will be downloaded:

On linux, can be something like ~/Download

On windows, can be something like: C:/Users/myusername/Downloads/protocol999 If Monkeys are running OK but no csv's are downloaded, make sure the docker username has write access to the folder

DEBUG

Show debug messages: FALSE / TRUE

screenshot

Should the monkey's take screenshots of each screen they see?

The images will be in outputs/screenshots

debug_file

Store the debug info in a file in outputs/debug

console_logs

Store console logs of browser? Logs are stored in outputs/logs

open_VNC

Show the info to open VNC viewer to see what the monkey's are up to

pid

Protocol id

uid_URL

Include user id in the protocol URL? If true, the uid are predefined

local_or_server

Run the protocol locally or on server c("local", "server) This variable will be inferred from whichever of local_folder_tasks or server_folder_tasks has information. See https://gorkang.github.io/jsPsychR-manual/qmd/04-jsPsychMonkeys.html#launch-monkeys-on-a-server for more information

local_folder_tasks

Local folder where the protocol is stored

server_folder_tasks

Location of the protocol in the server

disable_web_security

Run with CORS disabled? Needed for local protocols that load videos: FALSE / TRUE

initial_wait

Initial wait (in seconds) after entering the main protol page, before the monkeys start to complete the protocol. If the protocol has lots of images to pre-load should be a number big enough for them to download.

wait_retry

How many seconds to wait before retrying, after no elements are found in a page.

forced_random_wait

At some random point in the protocol, the monkey should wait for a random period of time? FALSE / TRUE

forced_refresh

Force a full refresh for some of the monkeys? This is useful to test if the monkey's can continue a protocol after they exit.

forced_seed

Random seed

Value

a list of parameters for the rest of the jsPsychMonkeys functions

Examples

try(set_parameters(parameters_monkeys = list(uid = 888, uid_URL = TRUE, forced_seed = 11,
                                       local_folder_tasks = "~/Downloads/protocol999",
                                       screenshot = FALSE, forced_refresh = FALSE,
                                       debug = TRUE, open_VNC = TRUE)))
#> $links_tar
#> $links_tar$times_repeat_protocol
#> [1] 1
#> 
#> $links_tar$time_to_sleep_before_repeating_protocol
#> [1] 1
#> 
#> 
#> $docker
#> $docker$browserName
#> [1] "chrome"
#> 
#> $docker$big_container
#> [1] FALSE
#> 
#> $docker$keep_alive
#> [1] FALSE
#> 
#> $docker$folder_downloads
#> [1] "/home/emrys/Downloads"
#> 
#> $docker$credentials_folder
#> [1] ".vault/"
#> 
#> 
#> $debug
#> $debug$DEBUG
#> [1] TRUE
#> 
#> $debug$screenshot
#> [1] FALSE
#> 
#> $debug$debug_file
#> [1] FALSE
#> 
#> $debug$console_logs
#> [1] FALSE
#> 
#> $debug$open_VNC
#> [1] TRUE
#> 
#> 
#> $remDr_params
#> $remDr_params$disable_web_security
#> [1] FALSE
#> 
#> 
#> $task_params
#> $task_params$pid
#>      [,1] 
#> [1,] "999"
#> 
#> $task_params$uid_URL
#> [1] TRUE
#> 
#> $task_params$local_or_server
#> [1] "local"
#> 
#> $task_params$local_folder_tasks
#> [1] "~/Downloads/protocol999"
#> 
#> $task_params$server_folder_tasks
#> NULL
#> 
#> $task_params$initial_wait
#> [1] 2
#> 
#> $task_params$wait_retry
#> [1] 0.1
#> 
#> $task_params$forced_random_wait
#> [1] FALSE
#> 
#> $task_params$forced_refresh
#> [1] FALSE
#> 
#> $task_params$forced_seed
#> [1] 11
#> 
#>